+++ lib/components/Layout/Footer.tsx
@@ -1,11 +1,11 @@
 // import { CSSObject } from '@emotion/react';
 // export const Footer = ({ children }) => {
 //   return (
-//     <footer css={baseStyles.footer}>Here's some footer stuff {children}</footer>
+//     <footer css={localStyles.footer}>Here's some footer stuff {children}</footer>
 //   );
 // };
 
-// const baseStyles: { [key: string]: CSSObject } = {
+// const localStyles: { [key: string]: CSSObject } = {
 //   footer: {
 //     minHeight: '100px'
 //   }