+++ lib/components/Layout/Main.tsx
@@ -1,22 +1,32 @@
-import { CSSObject } from '@emotion/react';
 import { ReactNode } from 'react';
+import { CSSObject } from '@emotion/react';
+import { baseColors } from '../../styling/baseTheme';
 
 type MainProps = {
   children?: ReactNode;
   className?: CSSObject;
+  isDark?: boolean;
+  backgroundImage?: string;
 };
 
-export const Main: React.FC<MainProps> = ({ children, className }) => {
+export const Main: React.FC<MainProps> = ({
+  children,
+  className,
+  isDark = false
+}) => {
+  // console.log(className);
   return (
-    <main className="pa-main" css={[baseStyles.main, className?.main]}>
+    <main className="pa-main" css={[localStyles(isDark).main, className?.main]}>
       {children}
     </main>
   );
 };
 
-const baseStyles = {
+const localStyles = (isDark: boolean) => ({
   main: {
     display: 'flex',
-    justifyContent: 'center'
+    justifyContent: 'center',
+    backgroundColor: isDark ? baseColors.primary[500] : baseColors.tertiary[50],
+    color: isDark ? baseColors.tertiary[50] : baseColors.primary[500]
   }
-};
+});