+++ lib/components/Layout/index.ts
@@ -2,5 +2,4 @@ export { Main } from './Main';
 export { InnerWidth } from './InnerWidth';
 export { Header } from './Header';
 export { Block } from './Block';
-export { TwoColumn } from './TwoColumns';
 export { Image } from './Image';