+++ lib/components/Lists/List.tsx
@@ -1,5 +1,4 @@
 import { ReactNode } from 'react';
-import { shadesAndTints } from '../../styling/baseTheme';
 import { CSSObject } from '@emotion/react';
 
 type ListProps = {
@@ -9,16 +8,15 @@ type ListProps = {
 
 export const List: React.FC<ListProps> = ({ className, children }) => {
   return (
-    <ul css={[baseStyles.ul, className?.ul]} className="pa-list">
+    <ul css={[localStyles.ul, className?.ul]} className="pa-list">
       {children}
     </ul>
   );
 };
 
-const baseStyles: CSSObject = {
+const localStyles: CSSObject = {
   ul: {
     borderRadius: '6px',
-    backgroundColor: shadesAndTints.tertiaryLight,
     fontWeight: '500',
     padding: '20px',
     margin: 0,