+++ lib/index.tsx
@@ -2,6 +2,7 @@ import './styling/normalizer.css';
 
 export * from './components';
 export * as PAHooks from './hooks';
+export * from './styling';
 export {
   hexToRgb,
   getLuminance,