+++ lib/styling/Globals.tsx
@@ -1,16 +1,33 @@
-/** @jsxImportSource @emotion/react */
 import { Global, css } from '@emotion/react';
-import { testTheme } from './baseTheme';
+import { baseTypography } from './baseTheme';
 
-const PhantomGlobals = () => (
+export const Globals = () => (
   <Global
-    styles={[
-      css`
-        @import url('./styling/normalizer.css');
-      `,
-      testTheme
-    ]}
+    styles={css`
+      @import url('./styling/normalizer.css');
+      h1 {
+        ${css(baseTypography.h1)}
+      }
+      h2 {
+        ${css(baseTypography.h2)}
+      }
+      h3 {
+        ${css(baseTypography.h3)}
+      }
+      h4 {
+        ${css(baseTypography.h4)}
+      }
+      h5 {
+        ${css(baseTypography.h5)}
+      }
+      h6 {
+        ${css(baseTypography.h6)}
+      }
+      a {
+        ${css(baseTypography.a)}
+      p {
+        ${css(baseTypography.body)}
+      }
+    `}
   />
 );
-
-export default PhantomGlobals;